pyproject.toml
[project]
name = "a2a-fight"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Nancycho", email = "nancy1xn1xn@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.10",
    "google-adk>=1.5.0",
    "uvicorn>=0.35.0",
]

[project.scripts]
a2a-fight = "a2a_fight:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
